Create LICENSE
Hallo hannahrobot,
i really appreciate your tutorial and code exmples. It was very helpful in understanding the principles of an online game. My friend is a teacher and we try now to implement our own version of this game for his school. Therefore I would like to ask you, if you may add a license to this project?